http: Move If-Modified-Since code in its own file
This code was picked from Golang source code, it's under a different
license, and ideally it should be synced with its upstream every now
and then.

For all these reasons, it better be in its own file.

For historical background:
- <https://github.com/etix/mirrorbits/pull/131>
- 75e0b5f0413e0af548f19e1e1eef1111549d1f18 and the few commits that
  follow
