+++ http/http.go
@@ -34,26 +34,6 @@ import (
 	"gopkg.in/tylerb/graceful.v1"
 )
 
-// condResult is the result of an HTTP request precondition check.
-// See https://tools.ietf.org/html/rfc7232 section 3.
-type condResult int
-
-const (
-	condNone condResult = iota
-	condTrue
-	condFalse
-)
-
-// TimeFormat is the time format to use when generating times in HTTP
-// headers. It is like [time.RFC1123] but hard-codes GMT as the time
-// zone. The time being formatted must be in UTC for Format to
-// generate the correct format.
-//
-// For parsing this time format, see [ParseTime].
-const TimeFormat = "Mon, 02 Jan 2006 15:04:05 GMT"
-
-var unixEpochTime = time.Unix(0, 0)
-
 var (
 	log = logging.MustGetLogger("main")
 )
@@ -238,86 +218,6 @@ func (h *HTTP) requestDispatcher(w http.ResponseWriter, r *http.Request) {
 		h.checksumHandler(w, r, ctx)
 	}
 }
-
-// The functions below were picked from go/src/net/http/fs.go
-//
-// Copyright 2009 The Go Authors. All rights reserved.
-//
-// Redistribution and use in source and binary forms, with or without
-// modification, are permitted provided that the following conditions are
-// met:
-// 
-//    * Redistributions of source code must retain the above copyright
-// notice, this list of conditions and the following disclaimer.
-//    * Redistributions in binary form must reproduce the above
-// copyright notice, this list of conditions and the following disclaimer
-// in the documentation and/or other materials provided with the
-// distribution.
-//    * Neither the name of Google Inc. nor the names of its
-// contributors may be used to endorse or promote products derived from
-// this software without specific prior written permission.
-// 
-// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-// "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-// LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
-// A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
-// OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
-// SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
-// LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
-// DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
-// THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-// (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-// OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-
-// isZeroTime reports whether t is obviously unspecified (either zero or Unix()=0).
-func isZeroTime(t time.Time) bool {
-	return t.IsZero() || t.Equal(unixEpochTime)
-}
-
-func setLastModified(w http.ResponseWriter, modtime time.Time) {
-        if !isZeroTime(modtime) {
-                w.Header().Set("Last-Modified", modtime.UTC().Format(TimeFormat))
-        }
-}
-
-func checkIfModifiedSince(r *http.Request, modtime time.Time) condResult {
-	if r.Method != "GET" && r.Method != "HEAD" {
-		return condNone
-	}
-	ims := r.Header.Get("If-Modified-Since")
-	if ims == "" || isZeroTime(modtime) {
-		return condNone
-	}
-	t, err := http.ParseTime(ims)
-	if err != nil {
-		return condNone
-	}
-	// The Last-Modified header truncates sub-second precision so
-	// the modtime needs to be truncated too.
-	modtime = modtime.Truncate(time.Second)
-	if modtime.Before(t) || modtime.Equal(t) {
-		return condFalse
-	}
-	return condTrue
-}
-
-func writeNotModified(w http.ResponseWriter) {
-	// RFC 7232 section 4.1:
-	// a sender SHOULD NOT generate representation metadata other than the
-	// above listed fields unless said metadata exists for the purpose of
-	// guiding cache updates (e.g., Last-Modified might be useful if the
-	// response does not have an ETag field).
-	h := w.Header()
-	delete(h, "Content-Type")
-	delete(h, "Content-Length")
-	delete(h, "Content-Encoding")
-	if h.Get("Etag") != "" {
-		delete(h, "Last-Modified")
-	}
-	w.WriteHeader(http.StatusNotModified)
-}
-
-// End of functions from go/src/net/http/fs.go
 
 func (h *HTTP) mirrorHandler(w http.ResponseWriter, r *http.Request, ctx *Context) {
 	//XXX it would be safer to recover in case of panic
